seachInput and filter components moved to containers folder
