+++ src/App.test.js
@@ -3,6 +3,6 @@ import App from './components/App';
 
 test('renders learn react link', () => {
   render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
+  const linkElement = screen.getByText(/CRYPTOS/i);
   expect(linkElement).toBeInTheDocument();
 });
